Split IO handling code into separate handler modules
Currently, we have a direct, explicit link between the VUART IO and the
"outbound" data, currently to stdio.

We'll want to do multiple things with the VUART data in future (eg.,
logging).

This change introduces "handlers"; a struct of callbacks that are
invoked when UART data is available, or when a (handler-provided) poll
descriptor returns new events. We convert the stdio code into a handler
as part of this.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>